unified invitation dialog
